+++ src/USA/MoMoe/MentionMe/Main.java
@@ -1,6 +1,5 @@
 package USA.MoMoe.MentionMe;
 
-import org.bukkit.ChatColor;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.plugin.java.JavaPlugin;
 
@@ -11,12 +10,22 @@ public class Main extends JavaPlugin {
     @Override
     public void onEnable() {
         // Default Config Options
-        config.addDefault("NotifyInChat", true);
-        config.addDefault("NotifyInActionBar", true);
-        config.addDefault("MentionColor", "&e");
-        config.addDefault("HashtagColor", "&d");
-        config.addDefault("MentionMessage", "&aYou were mentioned by %player%&a !");
-        config.addDefault("Sound", "ENTITY_EXPERIENCE_ORB_PICKUP");
+        config.options().header(
+                "Thank you for downloading MentionMe v0.2.0!\n" +
+                "The \"title-time\" option below is in server ticks. 20 = 1 second\n" +
+                "You can get a full list of sounds at: https://hub.spigotmc.org/javadocs/spigot/org/bukkit/Sound.html\n" +
+                "The MentionMe source code is available at: https://github.com/MoMoe0/MentionMe\n");
+        config.addDefault("notify-in-chat", true);
+        config.addDefault("notify-in-actionbar", true);
+        config.addDefault("mention-color", "&e");
+        config.addDefault("mention-message", "&aYou were mentioned by %player%&a !");
+        config.addDefault("notify-in-title", true);
+        config.addDefault("title-message", "&aMentioned by:");
+        config.addDefault("subtitle-message", "%player%");
+        config.addDefault("title-time", 50);
+        config.addDefault("enable-sound", true);
+        config.addDefault("sound", "ENTITY_EXPERIENCE_ORB_PICKUP");
+        config.addDefault("hashtag-color", "&d");
         config.options().copyDefaults(true);
         saveConfig();
 
@@ -26,7 +35,7 @@ public class Main extends JavaPlugin {
         // Register Commands
         this.getCommand("mentionmereload").setExecutor(new ReloadCommand(this));
         
-        getLogger().info(ChatColor.GREEN + "MentionMe Enabled!");
+        getLogger().info("MentionMe Enabled!");
     }
 
     @Override
@@ -35,6 +44,6 @@ public class Main extends JavaPlugin {
         reloadConfig();
         saveConfig();
 
-        getLogger().info(ChatColor.RED + "MentionMe Enabled!");
+        getLogger().info("MentionMe Disabled!");
     }
 }