+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Change Log
 ===
 
+###v0.2.0 (12/27/2016)
+**Added Features**:
+* Title & Subtitle support.
+* Made config more readable and added documentation (Please delete and reload the config).
+* Sound can now be toggled off.
+
+**Fixes**:
+1. Fixed the `onDisable` message.
+
+---
 
 ###v0.1.0 (12/27/2016)
 *Initial Commit*